+++ example/example/settings_docs.py
@@ -1,8 +1,0 @@
-# flake8: noqa
-import os
-import sys
-
-sys.path.insert(0, os.getcwd())
-sys.path.insert(0, os.path.join(os.getcwd(), os.pardir))
-# noinspection PyUnresolvedReferences
-from .settings import *