+++ getpaid/backends/dummy/__init__.py
@@ -1,3 +1,3 @@
-from .processor import PaymentProcessor
+from .processor import PaymentProcessor  # noqa
 
 default_app_config = 'getpaid.backends.dummy.apps.GetpaidDummyAppConfig'