+++ getpaid/processor.py
@@ -1,5 +1,8 @@
 from abc import abstractmethod, ABC
 
+from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
+
 
 class BaseProcessor(ABC):
     display_name = None
@@ -7,9 +10,13 @@ class BaseProcessor(ABC):
     logo_url = None
     slug = None  # for friendly urls
     method = 'GET'
+    template_name = None
 
     def __init__(self, payment):
         self.payment = payment
+        self.path = payment.backend
+        if self.slug is None:
+            self.slug = payment.backend  # no more Mr. Friendly :P
 
     def get_form(self, post_data):
         """
@@ -57,3 +64,18 @@ class BaseProcessor(ABC):
     @abstractmethod
     def get_redirect_url(self):
         return
+
+    def get_template_names(self, view=None):
+        getpaid_config = getattr(settings, 'GETPAID', {})
+        backend_config = getpaid_config.get('BACKENDS', {}).get(self.path, {})
+
+        template_name = backend_config.get('POST_TEMPLATE')
+        if template_name is None:
+            template_name = getpaid_config.get('POST_TEMPLATE')
+        if template_name is None:
+            template_name = self.template_name
+        if template_name is None and hasattr(view, 'get_template_names'):
+            return view.get_template_names()
+        if template_name is None:
+            raise ImproperlyConfigured("Couldn't determine template name!")
+        return [template_name]