+++ getpaid/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path, include, re_path
+from django.conf.urls import url, include
 
 from . import views
 from .registry import registry
@@ -6,9 +6,9 @@ from .registry import registry
 app_name = 'getpaid'
 
 urlpatterns = [
-    re_path(r'^new/(?P<currency>[A-Z]{3})/$', views.CreatePaymentView.as_view(), name='create-payment'),
-    path('success/<uuid:pk>/', views.SuccessView.as_view(), name='payment-success'),
-    path('failure/<uuid:pk>/', views.FailureView.as_view(), name='payment-failure'),
-    path('callback/<uuid:pk>/', views.CallbackView.as_view(), name='callback-detail'),
-    path('backends/', include(registry.urls)),
+    url(r'^new/(?P<currency>[A-Z]{3})/$', views.CreatePaymentView.as_view(), name='create-payment'),
+    url(r'^success/(?P<pk>[0-9a-f-]+)/$', views.SuccessView.as_view(), name='payment-success'),
+    url(r'^failure/(?P<pk>[0-9a-f-]+)/$', views.FailureView.as_view(), name='payment-failure'),
+    url(r'^callback/(?P<pk>[0-9a-f-]+)/$', views.CallbackView.as_view(), name='callback-detail'),
+    url(r'^backends/', include(registry.urls)),
 ]