+++ tests/urls.py
@@ -1,29 +1,9 @@
-"""example URL Configuration
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals, absolute_import
 
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/1.9/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.conf.urls import url, include
-    2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
-"""
 from django.conf.urls import url, include
-from django.contrib import admin
 
-from orders.views import HomeView, OrderView
-
-app_name = 'getpaid_example'
 
 urlpatterns = [
-    url(r'^admin/', admin.site.urls),
-
-    url(r'^$', HomeView.as_view(), name='home'),
-    url(r'^order/(?P<pk>\d+)/$', OrderView.as_view(), name='order_detail'),
-    url(r'', include('getpaid.urls', namespace='getpaid')),
+    url(r'^payments/', include('getpaid.urls', namespace='getpaid')),
 ]